Moved direction strategies to separate module; added fix for 'window is undefined'
