+++ lib/network/modules/CanvasRenderer.js
@@ -149,10 +149,10 @@ class CanvasRenderer {
     // This is not something that will happen in normal operation, but we still need
     // to take it into account.
     //
+    if (typeof window === 'undefined') return;  // Note: `if (window === undefined)` doesn't work
+
     var myWindow = window;  // Grab a reference to reduce the possibility that 'window' is reset
                             // while running this method.
-    if (myWindow === undefined) return;
-
     let timer;
 
     if (this.requiresTimeout === true) {
